// apps/calendar/job/src/form/types.ts
export interface FormConfig {
  username: string;
  password: string;
  domain: string;
  form: string;
}

export type Entrie = {
  DateUpdated: string;
  EntryId: string;
  Field1: string;
  Field2: string;
  Field3: string;
  Field4: string;
  Field5: string;
  Field6: string;
  Field7: string;
  Field8: string;
  Field9: string;
  Field10: string;
  Field11: string;
  Field12: string;
  Field13: string;
  Field14: string;
};